feat(memo): write n_refs_provided/n_refs_quoted_verbatim to memo_traceability.json

+++ src/irc/memo/pipeline.py
@@ -53,7 +53,7 @@ class MemoOutput:
     skeleton: str
     draft: str
     audit_notes: str
-    traceability: dict[str, float]
+    traceability: dict[str, int]
     prompt_tokens_total: int
     completion_tokens_total: int
